==> 3_supp_materialsIV.Rmd <==
---
title: "Supp. Materials IV for `Bursty, irregular speech input to children predicts vocabulary size'"
date: "`r format(Sys.time(), '%d %B %Y')`"
output: 
  bookdown::pdf_document2:
    keep_tex: true
indent: true
toc: false
header-includes:
 \usepackage{float}
 \floatplacement{figure}{H}
---

```{r setup, include=FALSE}
library('ggplot2')
library('bookdown')
library('kableExtra')
library('knitr')
library('dplyr')
library('tidyverse')
library('lmerTest')
library('lme4')
library('stargazer')

opts_chunk$set(echo = TRUE,warning=FALSE,error=FALSE,message=FALSE)
opts_knit$set(root.dir='/Users/megcychosz/Library/CloudStorage/Box-Box/box-group-lena-studies/its_dataframes/')
```

# Supp. Materials IV: quantifying relationships with sleep

Do children who take longer naps have larger vocabularies?

```{r, read in data}
romeo_demo <- read.csv('../noise_pilotData/everyday_speech/dataframes/romeo_demo.csv')
l2t_demo <- read.csv('../noise_pilotData/everyday_speech/dataframes/l2t_demo.csv')


pre_sleep_cds <- read.csv('../noise_pilotData/everyday_speech/dataframes/everyday_cds_ods.csv') %>%
  filter(corpus!='E' & corpus!='J' & corpus!='A') %>% # remove Jessica's kids and kids w CIs (E & A)
  filter(!child_id %in% c('679LTP1','679LTP2','665LTP1','665LTP2','608LTP1','608LTP2','605LTP2','605LTP1', '079LTP1')) # remove addtl kids w/ CIs 
```

```{r, effects of sleep times on vocabulary}
center_scale <- function(x) {
    scale(x, scale = FALSE)
}

pre_sleep_cds_romeo <- pre_sleep_cds %>%
  filter(corpus=='R') %>%
  mutate(Study='Romeo') %>%
  merge(., romeo_demo, by=c('child_id'))

pre_sleep_cds_l2t <- pre_sleep_cds %>%
  filter(corpus!='R') %>%
  merge(., l2t_demo, by=c('child_id'))

sleep_cds <- pre_sleep_cds_romeo %>% 
  rbind(pre_sleep_cds_l2t) %>%
  filter(total_hrs>5) %>%
  filter(child_id!='015LTP2') 

# quantify the minutes of sleep in each recording
sleep_cds_dur <- sleep_cds %>%
  filter(!sleep_pred=='0') %>% # three recordings have no sleep, so are removed
  group_by(child_id) %>%
  summarize(sleep_dur = sum(Duration_Secs)) %>%
  mutate(sleep_dur_mins = minute(seconds_to_period(sleep_dur)))
            
sleep_cds_model<- sleep_cds %>%
  distinct_at(., vars(child_id), .keep_all = T) %>%
  select(Female, maternal_education_level, age_mos, PPVT_GSV, ResearchID, child_id, date) %>%
  merge(., sleep_cds_dur, by='child_id') %>%
  filter(age_mos<42) %>%
  #mutate(sleep_dur_mins2 = if_else((sleep_dur_mins==5), 0, sleep_dur_mins)) %>% # if a child only has 1 epoch of sleep, we assume it's unlikely to be a true nap, or really much of a rest at all (the results are the same whether or not we take this step)
  #rename(sleep_dur_mins2=sleep_dur_mins) %>%
  mutate(mat_ed_centered=center_scale(maternal_education_level),
           age_mos_centered=center_scale(age_mos),
           sleep_dur_mins_centered=center_scale(sleep_dur_mins))


# remove recordings that had any component recorded overnight since we are interested in naps
sleep_cds_model2 <- sleep_cds %>%
  rename(start_date=date) %>%
  mutate(end_date=substr(endTimeStamp,1,8),
         end_date2=sub(" .*", "", end_date)) %>% # remove everything after the space
  distinct(child_id, .keep_all = T) %>%
  filter(end_date2==start_date) %>% # remove the rows where the dates don't match as this indicates that part of the recording took place while the child was sleeping
  select(child_id) %>% # list of recordings we'll be keeping
  merge(., sleep_cds_model, by='child_id') %>% 
  filter(sleep_dur_mins>0) # remove all 0-minute naps 
```

```{r, fit the model}
baseline_sleep <- lmer(PPVT_GSV~age_mos_centered+Female+mat_ed_centered+(1|ResearchID), data=sleep_cds_model2) 
sleep_m1 <- lmer(PPVT_GSV~age_mos_centered+Female+mat_ed_centered+sleep_dur_mins_centered+(1|ResearchID), data=sleep_cds_model2) 
sleep_test <- anova(sleep_m1, baseline_sleep) # significant improvement and effect of naps on vocab
sleep_sum <- summary(sleep_m1)

# descriptive stats about the duration of remaining sleep in those recordings
mean(sleep_cds_model2$sleep_dur_mins)
sd(sleep_cds_model2$sleep_dur_mins)
range(sleep_cds_model2$sleep_dur_mins)

# how many unique children?
num_child <- sleep_cds_model2 %>%
  distinct(ResearchID) %>%
  nrow()
```

Here we evaluate the relationship between nap duration and concurrent vocabulary size. We first removed recordings that did not contain any sleep (N=3). We could not reliably distinguish between potential overnight sleep and naptimes, so we then removed all recordings that had any component recorded overnight to be maximally conservative. We also limited the analysis to children under 42 months who we assumed would still be taking daily naps. This resulted in N=115 remaining recordings from N=108 children. 

To model the relationship between nap duration and concurrent vocabulary size, we fit a baseline linear mixed effects model to predict children's vocabulary size for the N=115 recordings. The baseline model included a random intercept for child and fixed effects of Child Age (in months, centered and scaled), Maternal Education (centered and scaled), and Child Gender (contrast coded). To this model we added an effect of Nap Duration (centered and scaled) which improved upon the baseline model ($\chi$^2=`r round(sleep_test$Chisq[2],2)`, p=.048). The model summary is listed in Table 1. There was a significant, positive relationship between nap duration and concurrent vocabulary ($\beta$=`r round(sleep_sum$coefficients[5,1],4)`, p=.03), suggesting that children who tend to take longer naps may have larger receptive vocabularies.

```{r, nap model summary }
nap_m <- lme4::lmer(PPVT_GSV~age_mos_centered+Female+mat_ed_centered+sleep_dur_mins_centered+(1|ResearchID), data=sleep_cds_model2)  


stargazer(nap_m,
            #column.labels = c("Burst Amplitude", "Burst Duration"),
            model.names=FALSE, 
            label = "nap-model",
            dep.var.caption = "", 
            dep.var.labels.include = FALSE,  
            type = "latex",
            star.cutoffs=c(0.05,0.01,0.001), 
            star.char = c("*", "**", "***"),
            title="The relationship between nap duration and concurrent vocabulary size",  
            digits = 2, 
            ci = TRUE, 
            style = "all",
            order=c(5,1,2,3,4),
            covariate.labels = c("Intercept", "Age (mos)", "Gender:Female", "Mat. Ed.", "Nap Duration (mins)"))
```
